requirements.txt
bcrypt>=4.0.0
cryptography>=42.0.0
yfinance>=0.2.36
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.26.0
pandas-ta>=0.3.14b
finnhub-python>=2.4.19
nltk>=3.8.1
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
APScheduler>=3.10.4,<4.0.0
